+++ src/main/java/com/ncell/wangcai/pojo/cns/entry/document/TextCell.java
@@ -1,0 +1,14 @@
+package com.ncell.wangcai.pojo.cns.entry.document;
+
+import com.ncell.wangcai.pojo.cns.main.base.Cell;
+
+/**
+ * 文字细胞
+ *
+ * @author anliwei
+ * @Date 2020/6/7 9:40
+ */
+public class TextCell extends Cell {
+
+
+}
