+++ src/main/java/com/ncell/wangcai/pojo/cns/entry/sound/SoundTissue.java
@@ -1,0 +1,10 @@
+package com.ncell.wangcai.pojo.cns.entry.sound;
+
+import com.ncell.wangcai.pojo.cns.main.base.Tissue;
+
+/**
+ * @author anliwei
+ * @create 2020/6/8 18:28
+ */
+public class SoundTissue extends Tissue {
+}
