+++ src/main/java/com/ncell/wangcai/pojo/cns/exit/document/TextAgent.java
@@ -1,0 +1,10 @@
+package com.ncell.wangcai.pojo.cns.exit.document;
+
+import com.ncell.wangcai.pojo.cns.main.base.Agent;
+
+/**
+ * @author anliwei
+ * @create 2020/6/8 18:26
+ */
+public class TextAgent extends Agent {
+}
