+++ src/main/java/com/ncell/wangcai/pojo/cns/exit/document/TextCell.java
@@ -1,0 +1,12 @@
+package com.ncell.wangcai.pojo.cns.exit.document;
+
+import com.ncell.wangcai.pojo.cns.main.base.Cell;
+
+/**
+ * @author anliwei
+ * @Data 2020/6/7 9:40
+ */
+public class TextCell extends Cell {
+
+
+}
