+++ src/main/java/com/ncell/wangcai/pojo/cns/exit/image/ImageCell.java
@@ -1,0 +1,10 @@
+package com.ncell.wangcai.pojo.cns.exit.image;
+
+import com.ncell.wangcai.pojo.cns.main.base.Cell;
+
+/**
+ * @author anliwei
+ * @create 2020/6/8 18:30
+ */
+public class ImageCell extends Cell {
+}
