+++ src/main/java/com/ncell/wangcai/pojo/cns/exit/sound/SoundCell.java
@@ -1,0 +1,10 @@
+package com.ncell.wangcai.pojo.cns.exit.sound;
+
+import com.ncell.wangcai.pojo.cns.main.base.Cell;
+
+/**
+ * @author anliwei
+ * @create 2020/6/8 18:27
+ */
+public class SoundCell extends Cell {
+}
