--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/document/TextAgent.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.document;
-
-import com.ncell.wangcai.pojo.cns.main.base.Agent;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:26
- */
-public class TextAgent extends Agent {
-}
