--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/document/TextCell.java
@@ -1,12 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.document;
-
-import com.ncell.wangcai.pojo.cns.main.base.Cell;
-
-/**
- * @author anliwei
- * @Data 2020/6/7 9:40
- */
-public class TextCell extends Cell {
-
-
-}
