--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/document/TextTissue.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.document;
-
-import com.ncell.wangcai.pojo.cns.main.base.Tissue;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:21
- */
-public class TextTissue extends Tissue {
-}
