--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/image/ImageAgent.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.image;
-
-import com.ncell.wangcai.pojo.cns.main.base.Agent;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:31
- */
-public class ImageAgent extends Agent {
-}
