--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/image/ImageTissue.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.image;
-
-import com.ncell.wangcai.pojo.cns.main.base.Tissue;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:30
- */
-public class ImageTissue extends Tissue {
-}
