--- src/main/java/com/ncell/wangcai/pojo/cns/inputConverter/sound/SoundCell.java
@@ -1,11 +1,0 @@
-package com.ncell.wangcai.pojo.cns.inputConverter.sound;
-
-import com.ncell.wangcai.pojo.cns.main.base.Cell;
-
-/**
- * 相当于人类大脑的语言中枢的听觉语言中枢
- * @author anliwei
- * @create 2020/6/8 18:27
- */
-public class SoundCell extends Cell {
-}
