+++ src/main/java/com/ncell/wangcai/pojo/cns/main/stem/Stem.java
@@ -121,79 +121,6 @@ public class Stem implements Serializable {
     long refractoryPeriod;
 
 
-    /**
-     * @update
-     * 2020年7月17日22:48:46
-     * 表示方位的属性，只适用于感知细胞，
-     * 比如眼睛，表示兴奋细胞所在的位置
-     *
-     * 比如耳朵的毛细胞至少分左右两大类 。
-     *
-
-     *
-     * 坐标原点是注意的焦点细胞，
-     *
-     * 如果没有设置注意细胞，则默认的是图像正中
-     *
-     * 如果整个视野是偶数，则使用稍小一点的为中点
-     */
-
-    int absoluteXLocation;
-    int absoluteYLocation;
-
-    /**
-     *
-     * @update
-     * 2020年7月17日23:08:18
-     *
-     * 仿照位置关系细胞
-     * 设置不同的精度
-     * 4精度将整个屏幕横向分为4个，整个视野分成16部分
-     * 8精度将整个屏幕横向分为8个，整个视野分成64部分
-     * 16精度将整个屏幕横向分为16个，整个视野分成256部分
-     *
-     */
-
-    int roughXLocation4;
-    int roughYLocation4;
-
-    int roughXLocation8;
-    int roughYLocation8;
-
-    int roughXLocation16;
-    int roughYLocation16;
-
-
-
-
-    /**
-     *
-     * @update
-     * 2020年7月17日23:22:54
-     * 使用极坐标标记方向
-     *
-     * 360度
-     * 分成12份，每一部分30度
-     * 或者分成24份，每一部分15度
-     *
-     */
-
-    int polarCoordinates12;
-
-    int polarCoordinates24;
-
-    /**
-     * @update
-     * 2020年7月17日23:25:22
-     * 相对距离
-     * 采用int来存储
-     * 8等级
-     * 16个等级，
-     */
-
-    int roughSpan8;
-
-    int roughSpan16;
