--- src/main/java/com/ncell/wangcai/pojo/cns/outputConverter/image/ImageCell.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.outputConverter.image;
-
-import com.ncell.wangcai.pojo.cns.main.base.Cell;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:30
- */
-public class ImageCell extends Cell {
-}
