--- src/main/java/com/ncell/wangcai/pojo/cns/outputConverter/sound/SoundAgent.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.outputConverter.sound;
-
-import com.ncell.wangcai.pojo.cns.main.base.Agent;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:29
- */
-public class SoundAgent extends Agent {
-}
