--- src/main/java/com/ncell/wangcai/pojo/cns/outputConverter/sound/SoundCell.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.outputConverter.sound;
-
-import com.ncell.wangcai.pojo.cns.main.base.Cell;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:27
- */
-public class SoundCell extends Cell {
-}
