--- src/main/java/com/ncell/wangcai/pojo/cns/outputConverter/sound/SoundTissue.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.pojo.cns.outputConverter.sound;
-
-import com.ncell.wangcai.pojo.cns.main.base.Tissue;
-
-/**
- * @author anliwei
- * @create 2020/6/8 18:28
- */
-public class SoundTissue extends Tissue {
-}
