+++ src/main/java/com/ncell/wangcai/service/cns/entry/ImageToCellConvertService.java
@@ -1,0 +1,9 @@
+package com.ncell.wangcai.service.cns.entry;
+
+/**
+ * @author anliwei
+ * @Data 2020/6/7 16:54
+ */
+public interface ImageToCellConvertService {
+    void convertImageToCell();
+}
