+++ src/main/java/com/ncell/wangcai/service/cns/exit/CellToDocumentConvertService.java
@@ -1,0 +1,8 @@
+package com.ncell.wangcai.service.cns.exit;
+
+/**
+ * @author anliwei
+ * @Date 2020/7/5 20:39
+ */
+public interface CellToDocumentConvertService {
+}
