+++ src/main/java/com/ncell/wangcai/service/cns/exit/CellToSoundConvertService.java
@@ -1,0 +1,8 @@
+package com.ncell.wangcai.service.cns.exit;
+
+/**
+ * @author anliwei
+ * @Date 2020/7/5 20:37
+ */
+public interface CellToSoundConvertService {
+}
