+++ src/main/java/com/ncell/wangcai/service/cns/exit/impl/CellToDocumentConvertServiceImpl.java
@@ -1,0 +1,10 @@
+package com.ncell.wangcai.service.cns.exit.impl;
+
+import com.ncell.wangcai.service.cns.exit.CellToDocumentConvertService;
+
+/**
+ * @author anliwei
+ * @Date 2020/7/5 20:39
+ */
+public class CellToDocumentConvertServiceImpl implements CellToDocumentConvertService {
+}
