--- src/main/java/com/ncell/wangcai/service/cns/inputConverter/ImageToCellConvertService.java
@@ -1,9 +1,0 @@
-package com.ncell.wangcai.service.cns.inputConverter;
-
-/**
- * @author anliwei
- * @Data 2020/6/7 16:54
- */
-public interface ImageToCellConvertService {
-    void convertImageToCell();
-}
