--- src/main/java/com/ncell/wangcai/service/cns/inputConverter/SoundToCellConvertService.java
@@ -1,12 +1,0 @@
-package com.ncell.wangcai.service.cns.inputConverter;
-
-/**
- * @author anliwei
- * @Data 2020/6/7 16:54
- */
-public interface SoundToCellConvertService {
-    /**
-     * 将声音信号转化为细胞
-     */
-    public void convertSoundToCellService();
-}
