--- src/main/java/com/ncell/wangcai/service/cns/main/psychology/Impl/AttentionServiceImpl.java
@@ -1,37 +1,0 @@
-package com.ncell.wangcai.service.cns.main.psychology.Impl;
-
-import com.ncell.wangcai.service.cns.main.psychology.AttentionService;
-
-/**
- *
- * 注意点，
- * 注意半径
- * 两个关键指标
- * @author anliwei
- * @Date 2020/7/13 23:22
- */
-public class AttentionServiceImpl implements AttentionService {
-    /**
-     * 在同级别中选择pojo，相当于选择的宽度
-     */
-    @Override
-    public void pickPojoInSameLevel() {
-
-    }
-
-    /**
-     * 在下级中选择pojo，相当于选择的深度
-     */
-    @Override
-    public void pickPojoInSubLevel() {
-
-    }
-
-    /**
-     * 调整筛选其他待发放pojo的速度
-     */
-    @Override
-    public void adjustSpeedOfPojoRunService() {
-
-    }
-}
