--- src/main/java/com/ncell/wangcai/service/cns/outputConverter/CellToDocumentConvertService.java
@@ -1,8 +1,0 @@
-package com.ncell.wangcai.service.cns.outputConverter;
-
-/**
- * @author anliwei
- * @Date 2020/7/5 20:39
- */
-public interface CellToDocumentConvertService {
-}
