--- src/main/java/com/ncell/wangcai/service/cns/outputConverter/CellToImageConvertService.java
@@ -1,8 +1,0 @@
-package com.ncell.wangcai.service.cns.outputConverter;
-
-/**
- * @author anliwei
- * @Date 2020/7/5 20:38
- */
-public interface CellToImageConvertService {
-}
