--- src/main/java/com/ncell/wangcai/service/cns/outputConverter/impl/CellToDocumentConvertServiceImpl.java
@@ -1,10 +1,0 @@
-package com.ncell.wangcai.service.cns.outputConverter.impl;
-
-import com.ncell.wangcai.service.cns.outputConverter.CellToDocumentConvertService;
-
-/**
- * @author anliwei
- * @Date 2020/7/5 20:39
- */
-public class CellToDocumentConvertServiceImpl implements CellToDocumentConvertService {
-}
