+++ src/main/java/com/ncell/wangcai/service/dogService/manager/impl/ManagerServiceImpl.java
@@ -38,8 +38,9 @@ public class ManagerServiceImpl implements ManagerService {
 
     InputIndicator inputIndicator;
     CnsIndicator cnsIndicator;
-    ServiceIndicator serviceIndicator;
     OutputIndicator outputIndicator;
+
+    ServiceIndicator serviceIndicator;
 
 
     /**
